% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_func2.R
\name{anova_crd}
\alias{anova_crd}
\title{Anova analysis for given AST204 problem
anova_crd gets the anova table for given dataset whose treatments are given
rowwise and replications will be columnwise}
\usage{
anova_crd(data, info = c("anova", "All"))
}
\arguments{
\item{info}{if u want all information type "All"}
}
\value{
dataframe
}
\description{
Anova analysis for given AST204 problem
anova_crd gets the anova table for given dataset whose treatments are given
rowwise and replications will be columnwise
}
\examples{
obs = c(25,35,21,37,40,52,39,58,27,34,20,31)
a = 3
b = 4
N = 12
mat = matrix(obs, nrow = a, ncol = b, byrow = T)
anova_crd(mat)
}
