\name{data.adjacent.mat}
\alias{data.adjacent.mat}
\docType{data}
\title{
data frame of adjacent provinces in Iran
}
\description{
This a matrix belonging to 31 provinces of Iran. The rownames of this matrix are matched with the codes of administrative names in Iran. In this matrix adjacent provinces specify with 1 and others with 0.
}
\usage{data("data.adjacent.mat")}
\format{
  A matrix of adjacent provinces with equal rows and columns
}

\examples{
data(data.adjacent.mat)
}
