% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{t_test}
\alias{t_test}
\title{T-test for two given treatments vector}
\usage{
t_test(y1, y2)
}
\arguments{
\item{y2}{}
}
\value{
dataframe
}
\description{
T-test for two given treatments vector
}
