% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{T-test for two given treatments vector}
\usage{
t_test(y1, y2)
}
\arguments{
\item{y1}{}

\item{y2}{}
}
\value{
a dataframe of pooled variance and value of test_statistics 
for given two treatments vector
}
\description{
T-test for two given treatments vector
}
